WIP: Refactor ace core modules into ace_core.py
